==> 01_HOME_data-processing.Rmd <==
---
title: "01_merging"
author: "Puvvula"
date: "2022-09-13"
output: pdf_document
---

```{r}
library(tidyverse)
#library(skimr)
library(haven)
library(janitor)
library(naniar)
#library(corrplot)
library(reshape2)
```

Filter outcome data by year of child visit
```{r}
################## OUTCOME ###############################
# In wide format WPPSI at 5-year visit
dt_outcome<- read_sas("/Users/jpuvvula/Documents/data/outcomes.sas7bdat")|> 
  filter(Visit == "M60")|>
  select(c(1,3,4,6)) |>
  rename(subject_id = SubjectID) |>
  drop_na() |>
  clean_names()

write_csv(dt_outcome, "/Users/jpuvvula/Documents/data/home_16w_imp/mirec_transf/wppsi-5yr_wide.csv")

#### ? REPLACING MISSING WPPSI WITH WISC
dt_outcome<- read_sas("/Users/jpuvvula/Documents/data/outcomes.sas7bdat")|> 
  filter(Visit == "M60" | Visit== "P3")|>
  select(c(1:4, 6:9)) |>
  rename(subject_id = SubjectID) |>
  clean_names()

out_wide <- dt_outcome |>
  pivot_wider(names_from = visit, values_from = -c(subject_id, visit)) |>
  clean_names()

out_wide$wppsi_fsiq <- ifelse(is.na(out_wide$wppsi_fsiq_m60), 
                                out_wide$wisc_fsiq_p3, out_wide$wppsi_fsiq_m60)

out_wide$wppsi_viq <- ifelse(is.na(out_wide$wppsi_viq_m60), 
                                out_wide$wisc_vci_p3, out_wide$wppsi_viq_m60)

out_wide$wppsi_piq <- ifelse(is.na(out_wide$wppsi_piq_m60), 
                                out_wide$wisc_pri_p3, out_wide$wppsi_piq_m60)

dat_out <- out_wide |>
  select(c(1,14:16)) |> drop_na()
write_csv(dat_out, "/Users/jpuvvula/Documents/data/home_16w_imp/mirec_transf/wppsi-wisc-suppl.csv")
```

Exposure data
```{r}
#reading SAS data
############### EXPOSURES ####################################

#BLOOD
bl_cd<- read_sas("/Users/jpuvvula/Documents/data/bcd.sas7bdat") |> clean_names()
bl_pb<- read_sas("/Users/jpuvvula/Documents/data/bpb.sas7bdat") |> clean_names()
thg <- read_sas("/Users/jpuvvula/Documents/data/thg.sas7bdat") |> clean_names()

# SERUM
############ OC Pest ###########
#remove observations if flg_lod is NA for OP,OC,PBDE
oc_pest<- read_sas("/Users/jpuvvula/Documents/data/ocpest.sas7bdat") |> 
  clean_names() 

#removing flg_lod = NA
labres<-filter(oc_pest, complete.cases(flg_lod_lp))
#summary stats
dat<- oc_pest |>
  group_by(analyte_grp, analyte_code, visit) |>
  summarise(tested=n(),
            bel_lod= sum(flg_lod_lp)) |>
  mutate(pct_lod= round((bel_lod/tested)*100,2))
####################
pbde<- read_sas("/Users/jpuvvula/Documents/data/pbde.sas7bdat") |> clean_names()

####################
pcb<- read_sas("/Users/jpuvvula/Documents/data/pcb.sas7bdat") |> clean_names()

######################
#pfas and ct doesnt have missing values flagging obs below LOD
pfas<- read_sas("/Users/jpuvvula/Documents/data/pfas.sas7bdat") |> clean_names()
ct <- read_sas("/Users/jpuvvula/Documents/data/ct.sas7bdat") |> clean_names()

# URINE
###########################################################
###########################################################
arse<- read_sas("/Users/jpuvvula/Documents/data/ars_i_cd.sas7bdat") |> clean_names()

arse$visit<-"16W" 

#check duplicates by subject per visit per analyte
arse_dat |> group_by(subject_id, visit) |> filter(n()>1)

###########################################################
###########################################################

opfr<- read_sas("/Users/jpuvvula/Documents/data/fr.sas7bdat") |> clean_names()
pheno<- read_sas("/Users/jpuvvula/Documents/data/phen.sas7bdat") |> clean_names()
bpa<- read_sas("/Users/jpuvvula/Documents/data/phthbpa.sas7bdat") |> clean_names()
pyre<- read_sas("/Users/jpuvvula/Documents/data/pyr.sas7bdat") |> clean_names()
op_pest<- read_sas("/Users/jpuvvula/Documents/data/op.sas7bdat") |> clean_names()
op_pest$flg_lod<- replace_na(op_pest$flg_lod, 1)
```

```{r}
dat<- pyre |>
  group_by(analyte_grp, analyte_code, visit) |>
  summarise(tested=n(),
            bel_lod= sum(flg_lod)) |>
  mutate(pct_lod= round((bel_lod/tested)*100,2))
```

Processing exposure data
```{r}
#For Cotinine only (take values from result or from result machine)
ct$Result <- ifelse(is.na(ct$Result), ct$Result_Machine, ct$Result)

#create average for multiple entries [replacing with average values]
#Use for PFAS and Cotinine only
dt_imp<- pfas |> 
  group_by(SubjectID, Visit, Analyte_Code) |>
  summarise_at(vars("Result"), mean)

#summaries for all variables
dt <- thg [c(1,2,5,6)] |>
  pivot_wider(names_from = c(Analyte_Code),
              values_from = Result)

dt_out <-dt |> 
  dplyr::group_by(Visit) |>
  skim()

write_csv(dt_out, "con.csv")
```

Specific gravity adjustment
```{r}
sg<- read_csv("/Users/jpuvvula/Documents/data/home_16w_imp/creat_sg.csv") |>
  filter(visit=="16W")

dt<- read_csv("/Users/jpuvvula/Documents/data/home_16w_imp/mecpp_16w.csv") 

dat<- inner_join(dt,sg, by="subject_id")

med_sg<- median(dat$specific_gravity)
dat<- dat |> 
  mutate(mecpp_sg = ((dat$result.res*(med_sg-1))/(dat$specific_gravity-1)))

dat<- dat[c(1,2,8)] |> rename(mecpp = result.res, subject_id = subject_id) |>
  pivot_longer(!subject_id, names_to = "analyte", values_to = "conc")

write_csv(dat, "/Users/jpuvvula/Documents/data/home_16w_imp/mirec_transf/dat/mecpp.csv")

```

wide to long lipid adjusted and whole blood samples
```{r}
dat<- read_csv("/Users/jpuvvula/Documents/data/home_16w_imp/pbde47_16w.csv") 
dat<- dat[c(1,2)] |> rename(pbde47 = result.res) |>
  pivot_longer(!subject_id, names_to = "analyte", values_to = "conc")

write_csv(dat, "/Users/jpuvvula/Documents/data/home_16w_imp/mirec_transf/pbde47.csv")
```

Comparing machine read vs imputed values below LOD
```{r}
ct <- read_sas("/Users/jpuvvula/Documents/data/ct.sas7bdat") |> 
  clean_names() |> 
  filter(visit =="16W", flg_lod == "1") |>
  select("subject_id", "result_machine")

ct_imp<- read_csv("/Users/jpuvvula/Documents/data/home_16w_imp/cotinine_16w.csv") |>
  select("subject_id", "result.res")

ct_compare<- left_join(ct, ct_imp, by="subject_id")

ggplot(ct_compare, aes(x=result_machine, y=result.res))+
  geom_point(size=1.5)+
  xlim(-0.001, 0.016)+
  ylim(-0.001, 0.016)+
  geom_rug()+
  geom_density2d(bins=5, color="grey20", linetype="dashed")+
  geom_smooth(linetype="dashed", color="gray60", fill="gray75")+
  labs(x="Machine read - Cotinine",
       y="Left truncated imputation - Cotinine",
       caption = "Generated using 109 subjects from HOME below LOD (0.015 mcg/mL)")+
  theme_bw()+ theme(panel.border = element_blank(), 
                    panel.grid.major = element_blank(),
                    panel.grid.minor = element_blank(), 
                    axis.line = element_line(colour = "black"))

ggsave("/Users/jpuvvula/Documents/misc_figs/cotinine_comparision.tiff", 
       width = 4,height = 4, dpi=200)
```

Merge all exposures to a csv
```{r}
file_names<- list.files(path= "/Users/jpuvvula/Documents/data/home_16w_imp/mirec_transf/dat",
                        recursive = T,
                        pattern = "\\.csv$",
                        full.names = T)

dat<- read_csv(file_names)

write_csv(dat, "/Users/jpuvvula/Documents/data/home_16w_imp/mirec_transf/home_exp_fin.csv")
```

Unmeasured exposure summary
```{r}
dat<- read_csv("/Users/jpuvvula/Documents/data/home_16w_imp/mirec_transf/home_exp_fin.csv") |>
  pivot_wider(names_from = analyte, values_from = conc)

# select sg adjusted variables
dat_exp<- dat |>
  select(subject_id, dde, pbde47, m_pb_sg, p_pb_sg, tcs_sg, bpa_sg,mbp_sg, mbzp_sg, mcpp_sg, mehhp_sg,
         mehp_sg, meohp_sg, mecpp_sg, mep_sg, mibp_sg, dep_sg, detp_sg, dmp_sg, dmtp_sg,
         cotinine, 
         pb, hg, dma_sg, pcb118, pcb138, pcb153, pcb180, pfhxs, pfoa, pfos, bcetp_sg,
         bdcpp_sg, dbup_sg, dphp_sg, b_pb_sg)

write_csv(dat_exp, "/Users/jpuvvula/Documents/data/home_16w_imp/mirec_transf/home_exp_wide.csv")

dat<- dat |> drop_na()

gg_miss_upset(dat, nsets=5, nintersects=NA)

```

Final outcome observations
```{r}
#outcome
dat_out<- read_csv("/Users/jpuvvula/Documents/data/home_16w_imp/mirec_transf/wppsi-wisc-suppl.csv")

#dat_out<- read_csv("/Users/jpuvvula/Documents/data/wppsi_wisc_repla.csv")

#analytic dataset
dat_exp_out<- full_join(dat_exp, dat_out, by="subject_id")

#missing data vis for WPPSI
gg_miss_upset(dat_exp_out, nsets=5, nintersects=NA)

write_csv(dat_exp_out, "/Users/jpuvvula/Documents/data/home_16w_imp/mirec_transf/wppsi-exp_wide.csv")

# For 14-chemical analysis
dat_wppsi<- dat_exp_out |>
  select(-c(22:36)) |>
  drop_na()

write_csv(dat_wppsi, "/Users/jpuvvula/Documents/data/home_16w_imp/mirec_transf/14_chem/wppsi-exp_wide.csv")  
```


```{r}
################# COVARIATES #############################
# In wide format
dt_cov <- read_sas("/Users/jpuvvula/Documents/data/covariates.sas7bdat") |>
  clean_names() |>
  select(c(1:3, 5, 7, 11, 14, 18, 16, 34, 50:55, 57, 59:64))

write_csv(dt_cov, "/Users/jpuvvula/Documents/data/home_16w_imp/mirec_transf/covar_wppsi.csv")

#extracting just education
df <- dt_cov |>
  select(c(1,16)) |>
  mutate(edu3 = case_when(
    Mom_Edu_Cd %in% 1:6 ~ 1,
    Mom_Edu_Cd %in% 7:8 ~ 2,
    Mom_Edu_Cd > 8 ~ 3,
    is.na(Mom_Edu_Cd) ~ NA_real_)) |>
  select(-c(Mom_Edu_Cd))

```


Join exposure, outcome and covariates
```{r}
cov <- read_csv("/Users/jpuvvula/Documents/data/home_16w_imp/mirec_transf/home-covar_mire_transf.csv") |>
  select(c(1:10)) |>
  clean_names()

dat_exp_out <- read_csv("E:/BBK17/pj/wppsi_analy_dat/home_export/wppsi-exp_wide.csv")

home_wppsi<- left_join(dat_exp_out, cov, by="subject_id")

################################################################################


gg_miss_upset(mirec_wppsi, nsets=5, nintersects=NA)

#replace missing values with median

#birth length missing (n=2)
home_wppsi <- home_wppsi |> mutate(b_length = ifelse(is.na(b_length),
                                                       median(b_length, na.rm = T),
                                                       b_length))


write_csv(home_wppsi, "E:/BBK17/pj/wppsi_exp_out_comb/home-mirec.csv")

dat_cov<- read_csv("/Users/jpuvvula/Documents/data/home_16w_imp/mirec_transf/covar_wppsi.csv") |>
  select(c(1:3, 6, 9, 10, 14:23))

gg_miss_upset(dat_cov, nsets=5, nintersects=NA)



```

adding parabens to existing data
```{r}
#butylparaben
b_pb<- read_csv("/Users/jpuvvula/Documents/data/imputed/bpb_16w.csv")

#methyl-paraben
m_pb<- read_csv("/Users/jpuvvula/Documents/data/imputed/mpb_16w.csv")

#propyl-paraben
p_pb<- read_csv("/Users/jpuvvula/Documents/data/imputed/ppb_16w.csv")

dat<- bind_rows(b_pb, m_pb, p_pb) |>
  pivot_wider(names_from = analyte, values_from = result.res) |>
  clean_names()

rm(list = c("p_pb", "m_pb", "b_pb"))

sg<- read_csv("/Users/jpuvvula/Documents/data/home_16w_imp/creat_sg.csv") |>
  filter(visit=="16W")

dat<- inner_join(dat,sg, by="subject_id")

med_sg<- median(dat$specific_gravity)
dat<- dat |> 
  mutate(p_pb_sg = ((dat$ppb*(med_sg-1))/(dat$specific_gravity-1)))

dat_exp<- dat[c(1,2,9:11)] |>
  rename(visit="visit.x")

write_csv(dat_exp, "/Users/jpuvvula/Documents/data/home_16w_imp/mirec_transf/dat/parabens.csv")
```

#add education to the full data
```{r}

```
