==> 02.x.2_study_specific_missing_summary.Rmd <==
---
title: "study_specific_missing_summary"
author: "Jagadeesh Puvvula"
date: "2023-04-19"
output: pdf_document
---

#collect home data for lt_lod percent
```{r}
home_path <- "~/Documents/data/"

# Use list.files() with recursive = TRUE to search for SAS files in the current directory and its subdirectories
file_list <- list.files(path=home_path, pattern = "\\.sas7bdat$", recursive = F, full.names = T)

# Filter out files with specific names
excluded_files <- c("outcomes.sas7bdat", "covariates.sas7bdat")
file_list <- file_list[!basename(file_list) %in% excluded_files]

# Use map_dfr() to load and bind all the files together, selecting variables 1, 2, 5, and 9
data <- map_dfr(file_list,  ~select(read_sas(.x), 1, 2, 5, contains("Flg"))) |>
  mutate_all(~replace_na(., 0)) |>
  mutate(flg_lod=Flg_LOD+Flg_LOD_lp) |>
  mutate_all(as.factor) |>
  select(-c(4:6))

#export to mirec server
write_csv(data, paste0(home_path, "home_lt_lod_pct.csv"))
```

+========================================================================+

```{r}
library(pacman)
pacman::p_load(tidyverse, janitor)
path <- "E:/BBK17/pj/data_2023apr/obs_w_outcome/"

```

```{r}
dat<- read_csv(paste0(path, "wppsi_f_analy.csv")) |>drop_na() |>filter(cohort=="2")
x<- dat[1]
remove(dat)
```

#home flod variable
```{r}
home<- read_csv("E:/BBK17/pj/data_2023apr/home_lt_lod_pct.csv") |>
  mutate_at(vars(2:4), as.factor) |>
  rename(subject_id = SubjectID)
```


# count observations - unmeasured and < LOD
```{r, for triclosan}
match_dataframes <- function(df1, df2) {
  # Extract first six characters of subject_id from df1
  df1$id <- substr(as.character(df1$subject_id), 1, 6)
  
  # Extract specimen_bar_code from df2
  df2$id <- substr(as.character(df2$specimen_bar_code), 1, 6)
  
  # Filter the observations from df2 that match with the id from df1
  df2_matched <- df2[df2$id %in% df1$id, ]
  
  # Convert log.flag variable to a factor
  df2_matched$log.flag <- as.factor(df2_matched$lod.flag)
  
  # Count the number of observations in each level of log.flag
  count <- table(df2_matched$log.flag)
  
  # Return the count
  return(count)
}

```

#function to count subjects with lt LOD values
```{r}
count_lt_results <- function(x, labres, sample_size) {
  # Extract first six characters of subject_id from x
  x$id <- substr(x$subject_id, 1, 6)
  
  # Extract first six characters of micro_aliquot_barcode from labres
  labres$id <- substr(labres$specimen_bar_code, 1, 6)
  
  # Filter labres observations where id matches with x
  labres_match <- labres[labres$id %in% x$id, ]
  
  # Count the number of observations with "<" symbol in test_result_raw_result
  labres_lt <- labres_match[grep("<", labres_match$tests_results_raw_result), ]
  count_lt <- table(labres_lt$tests_results_test_name)
  
  # Display the value with "<" symbol by each factor in test_result_test_name
  cat("Number of observations with '<' symbol:\n")
  for (test_name in names(count_lt)) {
    cat(paste(test_name, ": ", count_lt[test_name], " (", 
              round(count_lt[test_name]/sample_size*100, 2), "%)\n", sep=""))
    complete_lt <- labres_lt[labres_lt$tests_results_test_name == test_name & 
                            grep("<", labres_lt$tests_results_raw_result), ]
    cat(paste("Complete value with '<' symbol in", test_name, ":", 
              paste0(complete_lt$tests_results_raw_result, collapse=", "), "\n"))
  }
}

```

#lod percentages for mirec only
```{r}
count_lt_results(x=x, labres = labres,
                 sample_size= 478)
```


```{r}
home_summ<- home |>
  filter(Visit=="16W") |>
  group_by(Analyte_Code) |>
  mutate(subject_id=as.character(subject_id)) |>
  inner_join(x, by="subject_id") |>
  summarise(count = sum(flg_lod == "1"),
            percent = round((sum(flg_lod == "1") / n()) * 100,2))
```
