==> 02.x_home_mirec_merge.Rmd <==
---
title: "merge_home_mirec"
author: "Jagadeesh Puvvula"
date: "2023-01-31"
output: pdf_document
---

```{r}
library(pacman)
pacman::p_load(tidyverse, janitor, naniar)
```


#Add molar sum of DEHP and di-methyl/di-ethyl OP exposures - MIREC data
```{r}
dat<- read_csv("E://BBK17//pj//MIREC_exposure//mirec_full.csv")

#dehp = MEHHP (294.34) + MEHP (278.34) + MEOHP (292.33) +MECPP (308.33) | mol. wt from pubchem
avg_molw <- (294.34+278.34+292.33+308.33)/4
avg_mol_dm<- (126.05+141.11)/2
avg_mol_de<- (154.10+169.16)/2

dat <- dat |>
  mutate(
    dehp = (mehhp_sg / 294.34 * avg_molw) + (mehp_sg / 278.34 * avg_molw) +
      (meohp_sg / 292.33 * avg_molw) + (mecpp_sg / 308.33 * avg_molw),
    op_de = (dep_sg / 154.10 * avg_mol_de) + (detp_sg / 169.16 * avg_mol_de),
    op_dm = (dmp_sg / 126.05 * avg_mol_dm) + (dmtp_sg / 141.11 * avg_mol_dm)) |>
  select(-c(pbde153_lp, mehhp_sg, mehp_sg, meohp_sg, mecpp_sg, dep_sg, detp_sg, dmp_sg, dmtp_sg))

write_csv(dat, "E:/BBK17/pj/wppsi_exp_out_comb/wppsi_home_mirec.csv")
```


ADD parabens data
```{r}
final<- read_csv("E:/BBK17/pj/wppsi_exp_out_comb/wppsi_home_mirec.csv")


parabens<- read_csv("E:/BBK17/pj/add_parabens/parab_sg_all.csv") |>
  select(-c(2)) |>
  distinct(subject_id, .keep_all = T) # removed 50 duplicates with identical values


dat<- left_join(final, parabens, by="subject_id") |>
  drop_na()

# replacing the final analytical dataset with additional variables from paraben
write_csv(dat, "E:/BBK17/pj/wppsi_exp_out_comb/wppsi_home_mirec.csv")
```


#MIREC current data version 1933 * 48 
```{r}
write_csv(dat, "E:/BBK17/pj/data_2023apr/mirec_full.csv")

vis_miss(dat[-c(1)], cluster = T)

#filter data if any one outcome is present n=707
dat_n <- dat %>% 
  filter(rowSums(select(., 29:36), na.rm = TRUE) != 0)

vis_miss(dat_n[-c(1)], cluster = T)

write_csv(dat_n, "E:/BBK17/pj/data_2023apr/obs_w_outcome/mirec_w_out.csv")
```

#Load both HOME and MIREC data with at least 1 outcome and work from here
```{r}
path <- "E:/BBK17/pj/data_2023apr/obs_w_outcome/"

mirec<- read_csv(paste0(path, "mirec_w_out.csv")) |>
  clean_names() |>
  mutate_if(is.numeric, ~ round(., 3)) |>
  mutate(center = str_sub(subject_id, 1, 2), cohort= "2") |>
  rename_all(~ str_replace(., "(_sg|sg|_lp)", ""))|>
  rename(pfhxs=pf_hx_spl, pfoa=pfoapl, pfos=pfospl, basc_inz=basc_int,
         gest_age=gest_age_dv, wppsi_viq=viq, wppsi_piq=piq, wppsi_fsiq=fsiq)

home<- read_csv(paste0(path, "home_w_out.csv"))|>
  clean_names() |>
  mutate_if(is.numeric, ~ round(., 3)) |>
  mutate(center = "999", cohort= "1") |>
  rename_all(~ str_replace(., "(_sg|sg|_lp)", ""))|>
  rename(mom_age=mom_age_deliv, race=race_cd)
```


#check if variables present in both data
```{r}
home_vars <- setdiff(names(home), names(mirec))
cat("only available in HOME:", paste(home_vars, collapse = ", "))

mirec_vars <- setdiff(names(mirec), names(home))
cat("only available in MIREC:", paste(mirec_vars, collapse = ", "))
################################################################################

#sort the order of variables that match with mirec dataframe
common_vars <- intersect(names(mirec), names(home))
home_sorted <- home |>
  select(all_of(common_vars)) |>
  bind_cols(home |>
              select(-all_of(common_vars)))
```

#Join mirec and home_sorted dataframes [will be used for WPPSI, BRIEF and BASC]
```{r}
# Convert subject_id variable to character in home_sorted
home_sorted$subject_id <- as.character(home_sorted$subject_id)

# Combine mirec and home_sorted
combined_df <- bind_rows(mirec, home_sorted[1:50])

write_csv(combined_df, paste0(path, "home_mirec_w_out.csv"))
```

#missing data for all 1st trim outcomes
```{r}
vis_miss(combined_df[-c(1,49,50)], cluster = T)
```
