==> 02_MIREC_data-processing.Rmd <==
---
title: "mirec_data"
author: "puvvula"
date: "2022-10-05"
output: pdf_document
---

```{r}
library(pacman)
pacman::p_load(tidyverse, janitor, readxl, skimr, naniar)
```

Outcome: child 3 Yr scores
```{r}
#parent: SRS-2
#child: NEPSY, and WPPSI-3
#get WPPSI composite scores from here
################################################################################
################################################################################
wppsi<-  read_csv("E:/BBK17/4. BBK17_CD+ CRF data corrections_Nov18 2022/CDPLUS.2_NEURO.9_20221117/NDEV_WPPSI_Yes to NEURO BBK.csv") |>
  clean_names()|>
  select(centre, subject_id, age_testing_yrs, wppsi6_1, wppsi7_1, wppsi8_1) |>
  drop_na() |>
  rename(age= age_testing_yrs,
         viq= wppsi6_1,
         piq= wppsi7_1,
         fsiq= wppsi8_1)

#filter composite scores
write_csv(wppsi, "E://BBK17//pj//outcome//wppsi.csv")

################################################################################
################################################################################

basc<- read_csv("E:/BBK17/4. BBK17_CD+ CRF data corrections_Nov18 2022/CDPLUS.2_NEURO.3_20221117/MSAQ_DATA_BASC_Yes to NEURO BBK.csv") |>
  clean_names() |>
  select(1,2,168:170)|>
  rename(basc_ext=bastexcddv,
         basc_int=bastincddv,
         basc_bsi=bastbscddv) |>
  drop_na()


brief<- read_csv("E:/BBK17/4. BBK17_CD+ CRF data corrections_Nov18 2022/CDPLUS.2_NEURO.4_20221117/MSAQ_DATA_BRIEF_Yes to NEURO BBK.csv")|>
  clean_names() |>
  select(1,2,38,40)|>
  rename(brief_wm=brietwmcddv,
         brief_po=brietpocddv) |>
  drop_na()

basc_brief<- full_join(basc,brief[c(2:4)], by="subject_id")


write_csv(basc_brief, "E://BBK17//pj//outcome//basc_brief.csv")

################################################################################
################################################################################
```


First trimester - Exposure
```{r}

# triclosan (removing not reported records)
dat<- read_csv2("E://BBK17//3. BBK17_lab test correction on Dacima_Nov7 2022//MIREC_RESEARCHERS RESULTS table_code 001//QueryData_code001.csv") |>
  clean_names() |>
  rename(tests_results_raw_result =researcher_results_raw_result) |>
  filter(complete.cases(researcher_results_unit_set))

dat$tests_results_test_name<- str_extract(dat$researcher_results_test_name, "^.{3}") 
dat$tests_results_test_name<-as.factor(dat$tests_results_test_name)
dat$tests_results_raw_result<-as.numeric(dat$tests_results_raw_result)
dat<- dat |> filter(as.factor(tests_results_test_name) == "TRI")

# sum of triclosan glusuronide, free and sulfate
labres <- dat |> 
  group_by(specimen_bar_code) |>
  summarise(test.result=sum(tests_results_raw_result)) |>
  mutate(tests_results_test_name ="triclosan", visit = "12w")

#flag LT lod values for total triclosan 
# LOD = 0.12; from https://ehp.niehs.nih.gov/doi/epdf/10.1289/ehp.1408187
labres$lod.flag<-  as.numeric(ifelse(labres$test.result< 0.1200, 1, 0))

#Phthalates
labres<- read_csv2("E://BBK17/2. BBK17_pending lab tests_July19 2022//LAB//MIREC_MICROALIQOUT RESULTS table code 001-01.001-03//QueryData_code 001-01.csv") |>
  clean_names()|>
  rename(tests_results_test_name = micro_aliquot_tests_results_test_name,
         tests_results_raw_result =micro_aliquot_tests_results_raw_result)

#BPA, Urine Arsenic, OP pesticides, OPFR, Specific gravity
labres <-read_csv2("E://BBK17//1. BBK17_June10 2022//LAB//MIREC_TEST RESULTS table_codes 001.002.009.013.016.049.054.060.096.097.101//QueryData_code001.csv") |>
  clean_names() |> 
  mutate(id_join = str_sub(specimen_bar_code, 1,6))

df_1a <- df_1[c(5,6,10)] |> 
  pivot_wider(names_from = c(tests_results_test_name),
              values_from = tests_results_raw_result) 

num<- names(df_1a[c(2:15)])
df_1a[num]<- sapply(df_1a[num], as.numeric)

write_csv(df_1a, "E://BBK17//pj//pj_dat//ex_dat6.csv")

#Urine: Phthalates
df_2 <-read_csv2("E://BBK17//1. BBK17_June10 2022//LAB//MIREC_TEST RESULTS table_codes 001.002.009.013.016.049.054.060.096.097.101//QueryData_code002.csv") |>
  clean_names()

#Urine: Creatinine
df_6 <-read_csv2("E://BBK17//1. BBK17_June10 2022//LAB//MIREC_TEST RESULTS table_codes 001.002.009.013.016.049.054.060.096.097.101//QueryData_code049.csv") |> clean_names()

#Plasma: PBDE, PCB, OC Pesticides
df_3 <-read_csv2("E://BBK17//1. BBK17_June10 2022//LAB//MIREC_TEST RESULTS table_codes 001.002.009.013.016.049.054.060.096.097.101//QueryData_code009.csv") |> clean_names()

#Plasma: Tobacco and PFCs
df_4 <-read_csv2("E://BBK17//1. BBK17_June10 2022//LAB//MIREC_TEST RESULTS table_codes 001.002.009.013.016.049.054.060.096.097.101//QueryData_code013.csv") |> clean_names()

#Whole Blood: - metals
df_5 <-read_csv2("E://BBK17//1. BBK17_June10 2022//LAB//MIREC_TEST RESULTS table_codes 001.002.009.013.016.049.054.060.096.097.101//QueryData_code016.csv") |> clean_names()


#OPFR
labres<- read_csv2("E://BBK17/2. BBK17_pending lab tests_July19 2022//LAB//MIREC_MICROALIQOUT RESULTS table code 001-01.001-03//QueryData_code001-03.csv") |>
    clean_names()|>
    rename(tests_results_test_name = micro_aliquot_tests_results_test_name,
           tests_results_raw_result =micro_aliquot_tests_results_raw_result)

#Phthalate - MECPP
labres<- read_csv2("E://BBK17//2. BBK17_pending lab tests_July19 2022//LAB//MIREC_MICROALIQOUT RESULTS table code 001-01.001-03//QueryData_code 001-01.csv") |>
    clean_names()|>
    rename(tests_results_test_name = micro_aliquot_tests_results_test_name,
           tests_results_raw_result =micro_aliquot_tests_results_raw_result) |>
  filter(tests_results_test_name =="MECPP-M2")

```

Standardizing units
```{r}
#Metals: Lead & Mercury (mol. wt from pubchem)

#molecular weight x conversion factor to get values in mcg/L
pb<- 207.2/1e6 # micromole/L -> mcg/L
hg<-200.59/1e9 #nanomole/L -> mcg/L
#multiply above values with absolute result from imputed dataset
dt<- read_csv("E://BBK17/pj//imputed_data//HGSG.csv")
dt$result.res<-dt$result.res*hg*1e6

write_csv(dt, "E://BBK17/pj//imputed_unit_corr//HGSG.csv")

# PCBs,PBDE, DDE (mcg/L -> ng/L)
dt<- read_csv("E://BBK17/pj//imputed_data//BDE47PL.csv")
dt$result.res <- dt$result.res*1000
write_csv(dt, "E://BBK17/pj//imputed_unit_corr//BDE47PL.csv")

# DMAA micromol/L -> PPB (mol. wt from pubchem)
dt<- read_csv("E://BBK17/pj//imputed_data//DMAAUR.csv")
dt$result.res1 <- (dt$result.res*170)
write_csv(dt, "E://BBK17/pj//imputed_unit_corr//DMA.csv")

```

specific gravity standardization = ((analyte concentration x (median urine specific grav. -1)) / (1-individual specific grav. -1))
Creatinine adjustment = analyte concentration / creatinine concentration
lipid adjustment = (analyte concentration / lipid level)*102.6

```{r}
urine_sg<- read_csv("E://BBK17/pj//imputed_unit_corr//DENS.csv")
urine_sg$join_id<- str_extract(urine_sg$subject_id,  "^.{8}")

dt<- read_csv("E://BBK17/pj//imputed_unit_corr//MECPP.csv")
dt$join_id<- str_extract(dt$subject_id,  "^.{8}")
dat<- inner_join(dt,urine_sg, by="join_id")

med_sg<- median(dat$result.res.y)
dat$mecpp_sg <- ((dat$result.res.x*(med_sg-1))/(dat$result.res.y-1))

dat<- dat[c(1,2,10)] |> rename(mecpp = result.res.x, subject_id = subject_id.x) |>
  pivot_longer(!subject_id, names_to = "analyte", values_to = "conc")

write_csv(dat, "E://BBK17/pj//imputed_unit_corr//sg_adj//mecpp.csv")


################ creatinine adjustment ########################

cre<- read_csv("E://BBK17/pj//imputed_data//CRE.csv")
cre$join_id<- str_extract(cre$subject_id,  "^.{8}")

dt<- read_csv("E://BBK17/pj//imputed_data//COTISE.csv") 
dt$join_id<- str_extract(dt$subject_id,  "^.{8}")
dat<- inner_join(dt,cre, by="join_id")

dat$cotinine_cr <- (dat$result.res.x/dat$result.res.y)
dat<- dat[c(1,2,10)] |> rename(cotinine = result.res.x) |>
  pivot_longer(!subject_id.x, names_to = "analyte", values_to = "conc")

write_csv(dat, "E://BBK17/pj//imputed_data//cre_adj//cotinine.csv")

############ lipid adjustment for DDE, PBDE, PCB ############

lipi<-  read_csv("E://BBK17/pj//imputed_unit_corr//LIPIPL.csv") 
lipi$join_id<- str_extract(lipi$subject_id,  "^.{8}") 

dt <- read_csv("E://BBK17/pj//imputed_unit_corr//BDE47PL.csv") 
dt$join_id<- str_extract(dt$subject_id,  "^.{8}")
dat<- inner_join(dt,lipi, by="join_id")
 
dat$pbde47_lp <- (dat$result.res.x/dat$result.res.y)
dat<- dat[c(1,2,10)] |> rename(pbde47 = result.res.x, subject_id = subject_id.x) |>
  pivot_longer(!subject_id, names_to = "analyte", values_to = "conc")

write_csv(dat, "E://BBK17/pj//imputed_unit_corr//lipi_adj//pbde47.csv")

```

Converting all exposures to long format - join
```{r}
dat<- read_csv("E://BBK17/pj//imputed_data//HGSG.csv") 
dat<- dat[c(1,2)] |> rename(hg = result.res) |>
  pivot_longer(!subject_id, names_to = "analyte", values_to = "conc")

write_csv(dat, "E://BBK17/pj//imputed_data//merging//hg.csv")
```

Merging exposures to long format
```{r}
file_names<- list.files(path= "E://BBK17//pj//imputed_unit_corr//mirec_fin_merge",
                        recursive = T,
                        pattern = "\\.csv$",
                        full.names = T)

dat<- read_csv(file_names)
dat$subject_id<- str_extract(dat$subject_id,  "^.{6}")

write_csv(dat, "E://BBK17//pj//MIREC_exposure//mirec_exposure_long.csv")


############################################################
#Long to wide - exposures

dat<- read_csv("E://BBK17//pj//MIREC_exposure//mirec_exposure_long.csv")

dat_exp<- dat |>
  pivot_wider(id_cols=subject_id, names_from = analyte, values_from = conc,
              values_fn = list(conc= mean))


# variable 18 for DMA
dat_exp2<- dat_exp |>
  select(1,3,5,7,8,11,13,15,17,19,21,23,25,26,28,30,32,34,36,38,40,42,44,45,46,
         47,49,51,52,54,56,58,60,61,62,63,65)

write_csv(dat_exp2, "E://BBK17//pj//MIREC_exposure//mirec_exposure_wide.csv")

#unmeasured exposure summary
gg_miss_upset(dat, nsets=5, nintersects=NA)

#final exposure data
dat_exp2<- dat_exp2 |> drop_na()

write_csv(dat_exp2, "E://BBK17//pj//wppsi_analy_dat/dat_exp_mirec.csv")
```

all outcomes
```{r}
out<- read_csv("E://BBK17//pj//mirec_out.csv") |>
  group_by(subject_id) |>
  summarize(across(everything(), mean, na.rm = TRUE))

#find if there are duplicates
duplicated_ids <- out[duplicated(out$subject_id), "subject_id"]
```

Join all exposure and outcomes
```{r}
dat<- full_join(dat_exp2, out, by="subject_id")

write_csv(dat, "E://BBK17//pj//MIREC_exposure//mirec_exp-out_wide.csv")
```


join exp_out with covariates
```{r}
dt_cov<- read_csv("E://BBK17//pj//covariates//mirec_cov_fin.csv") 


dat_fin<- left_join(dat, dt_cov, by="subject_id")

write_csv(dat_fin, "E://BBK17//pj//MIREC_exposure//mirec_full.csv")


```



#################################################################################
#################################################################################
FOR WPPSI ANALYSIS
```{r}
dat_exp_out<- read_csv("E:/BBK17/pj//wppsi_analy_dat/exp_out-mirec.csv")
```


Join MIREC exposure outcome and covariates
```{r}
cov<- read_csv("E:/BBK17/pj//covariates/mirec_cov_fin.csv") |>
  select(c(1:10)) |> clean_names()

exp_out<- read_csv("E:/BBK17/pj//wppsi_analy_dat/exp_out-mirec.csv")


mirec_wppsi <- left_join(exp_out, cov, by="subject_id")
gg_miss_upset(mirec_wppsi, nsets=5, nintersects=NA)

#replace missing values with median

#HOME missing (n=16)
mirec_wppsi <- mirec_wppsi |> mutate(home_score_total = ifelse(is.na(home_score_total),
                                                               median(home_score_total, na.rm = T),
                                                               home_score_total))

#birth length missing (n=34)
mirec_wppsi <- mirec_wppsi |> mutate(b_length = ifelse(is.na(b_length),
                                                       median(b_length, na.rm = T),
                                                       b_length))

#birth weight missing (n=3)
mirec_wppsi <- mirec_wppsi |> mutate(b_wght = ifelse(is.na(b_wght),
                                                     median(b_wght, na.rm = T),
                                                     b_wght))

#mom education (n=2)
mirec_wppsi <- mirec_wppsi |> mutate(mom_edu = ifelse(is.na(mom_edu), 
                                                      4, mom_edu))

write_csv(mirec_wppsi, "E:/BBK17/pj/wppsi_exp_out_comb/wppsi-mirec.csv")
```


ADD paraben data
```{r}
dat<- read_csv("E:/BBK17/pj/add_parabens/mirec_parab_work/paraben_raw.csv") |>
  select(c(2, 8, 10, 12))|>
  clean_names()

#convert to long format for imputation
labres <- dat |>
  rename(specimen_bar_code="client_reference_number") |>
  pivot_longer(!specimen_bar_code, names_to = "tests_results_test_name",
               values_to = "tests_results_raw_result")

#########################################################
#########################################################
#read data for urine specific gravity correction
dt<- read_csv("E://BBK17/pj//imputed_unit_corr//p_pb.csv")

urine_sg<- read_csv("E://BBK17/pj//imputed_unit_corr//DENS.csv")
urine_sg$join_id<- str_extract(urine_sg$subject_id,  "^.{8}")

dt$join_id<- str_extract(dt$subject_id,  "^.{8}")
dat<- full_join(dt,urine_sg, by="join_id") |>
  drop_na()

med_sg<- median(dat$result.res.y)
dat$p_pb_sg <- ((dat$result.res.x*(med_sg-1))/(dat$result.res.y-1))

p_pb<- dat[c(1,10)] |> 
  rename(subject_id = subject_id.x) |>
  drop_na()

write_csv(p_pb, "E://BBK17/pj//imputed_unit_corr//sg_adj//p_pb.csv")

parab<- bind_cols(b_pb, m_pb, p_pb) |>
  select(c(1, 2, 4, 6)) |>
  rename(subject_id="subject_id...1")|>
  mutate(subject_id= str_extract(subject_id,  "^.{8}"),
         visit=="12w")

write_csv(parab, "E:/BBK17/pj/add_parabens/mirec_paraben_sg.csv")



################# COMBINE MIREC and HOME parabens
parab_home<- read_csv("E:/BBK17/pj/add_parabens/HOME_paraben_sg.csv")
parab_home$subject_id<- as.character(parab_home$subject_id)

parab_full<- bind_rows(parab, parab_home)
write_csv(parab_full, "E:/BBK17/pj/add_parabens/parab_sg_all.csv")
```

#April 19, 2024- generating paraben summary for aimin
```{r}
paraben_sept<- read_excel("E:\\BBK17\\5. BBK17_Lab results code 001-05_Feb7 2023\\Merged file from Ammanie_Sept1 2023\\BBK17_Parabens_For exploration (Ammanie).xlsx") |>
  select(c(1, 8:13, 15:17)) |>
  mutate_all(~if_else(. == "No result" | grepl("^<", .), NA, .)) |>
  mutate_all(as.numeric) |>
  pivot_longer(cols = everything(), 
               names_to = "variable", 
               values_to = "value")

summary_stats <- x |>
  group_by(variable) |>
  summarise(
    count_lt_LOD = sum(is.na(value)),
    count_detected = sum(!is.na(value)),
    min = sprintf("%.2f", min(value, na.rm = TRUE)),
    max = sprintf("%.2f", max(value, na.rm = TRUE)),
    `25th percentile` = sprintf("%.2f", quantile(value, 0.25, na.rm = TRUE)),
    `50th percentile` = sprintf("%.2f", quantile(value, 0.50, na.rm = TRUE)),
    `75th percentile` = sprintf("%.2f", quantile(value, 0.75, na.rm = TRUE)),
    average = sprintf("%.2f", mean(value, na.rm = TRUE))
  )

write_csv(summary_stats, "E:/BBK17/pj/paraben_summary.csv")
```
