==> 03.1_gam_res.Rmd <==
---
title: "GAM-results"
author: "Jagadeesh Puvvula"
date: "2023-02-02"
output: pdf_document
---

```{r}
library(tidyverse)
library(janitor)
#library(skimr)
library(modelsummary)
library(mgcv)
```


load exp and outcome data (n=617)
```{r, echo=FALSE, message=FALSE, results='hide'}
dat<- read_csv("E:/BBK17/pj/wppsi_exp_out_comb/wppsi_fin.csv") |>
  mutate_if(is.numeric, round, 4) |>
  drop_na() |>
  clean_names()|>
  select(-c(race, mom_edu))

cat<- c("cohort", "subject_id", "city", "age", "sex", "race_bin", "parity_n", "mom_edu_cat")
dat[cat]<- lapply(dat[cat], factor)

#Create race binary variable
#dat$race_bin<- as.factor(with(dat, ifelse(race=="white", "white", "non_white")))

summary(dat)
```

Pre-process data
```{r, echo=FALSE, message=FALSE}
# exposure and outcome variables
numeric_vars <- sapply(dat, is.numeric)

# Convert numeric variables to log2 scale
dat[c(5:7, 9:37)] <- log10(dat[c(5:7, 9:37)])
```

Generalized additive model: assuming Gaussian family - Identity link
CHANGE OUTCOME at each run
```{r, echo=FALSE, message=FALSE}
# Define predictors to be used in the loop
predictors <- list("s(bcetp_sg, by=cohort)", "s(bdcpp_sg, by=cohort)", "s(bpa_sg, by=cohort)", "s(dbup_sg, by=cohort)", 
                "s(dde_lp, by=cohort)", "s(dma_sg, by=cohort)", "s(dphp_sg, by=cohort)", "s(hgsg, by=cohort)", 
                "s(mbp_sg, by=cohort)", "s(mbzp_sg, by=cohort)", "s(mcpp_sg, by=cohort)", "s(mep_sg, by=cohort)",
                "s(mibp_sg, by=cohort)", "s(pbsg, by=cohort)", "s(pbde47_lp, by=cohort)","s(pcb118_lp, by=cohort)",
                "s(pcb138_lp, by=cohort)", "s(pcb153_lp, by=cohort)", "s(pcb180_lp, by=cohort)", "s(pf_hx_spl, by=cohort)",
                "s(pfoapl, by=cohort)", "s(pfospl, by=cohort)", "s(tcs_sg, by=cohort)", "s(dehp, by=cohort)", 
                "s(op_de, by=cohort)", "s(op_dm, by=cohort)","s(b_pb, by=cohort)","s(m_pb, by=cohort)", 
                "s(p_pb, by=cohort)")

covars<- ("+sex+mom_edu_cat+parity_n+mom_age+home_score_total+race_bin+cotinine")

results <- list()
plots<-list()

for (predictor in predictors) {
  formula <- as.formula(paste("wppsi_piq ~", predictor, covars))
  model <- gam(formula,
               method="GCV.Cp",
               data = dat)
  results[[predictor]] <- model
  plot <- plot.gam(model, pages = 1, shade = T, shade.col = "gray90")
  plot$title <- paste("Exposure - ", predictor)
  plots[[predictor]] <- plot
}

results

modelsummary(results, output = "E:/BBK17/pj/wppsi_results/gam_results-piq.csv")

save(results, file = "E:/BBK17/pj/wppsi_results/gam_res_piq.rda")
```
