==> 03.2_Variable_selection.Rmd <==
---
title: "03.2_variable selection"
author: "Jagadeesh Puvvula"
date: "2023-02-21"
output: pdf_document
---


```{r}
library(pacman)
pacman::p_load(tidyverse, reshape2, janitor)
path <- "E:/BBK17/pj/data_2023apr/obs_w_outcome/"
res<- "E:/BBK17/pj/data_2023apr/results/iq_paper/"
```


```{r}
dat <- read_csv(paste0(path, "wppsi_f_analy.csv")) |> drop_na() |>
  filter( cohort == "1" | cohort == "2" ) |> # 1- for HOME
  mutate_if(is.numeric, round, 4)|>
  clean_names() |>
  rename(sigma_DEHP=dehp, di_Ethyl_OP=op_de, di_Methyl_OP=op_dm, city=center)|>
  rename_at(vars(4:32), toupper) |>
  mutate(cohort = ifelse(cohort == "1", "HOME", "MIREC")) |>
  select(c(PB, HG, DMA, 
           DDE, PBDE47, PCB118, PCB138, PCB153, PCB180,
           PFHXS, PFOA, PFOS, 
           BCETP, BDCPP, DBUP, DPHP, 
           B_PB, M_PB, P_PB,
           TCS, BPA, 
           MBP, MBZP, MCPP, SIGMA_DEHP, MEP, MIBP,
           DI_ETHYL_OP, DI_METHYL_OP,
           cotinine, mom_age, home_score_total,
           wppsi_fsiq, wppsi_viq, wppsi_piq))
  
```


```{r, echo=FALSE, message=FALSE}
#Pre-process data
# convert exposures to log and center scale both x and y variables
dat <- dat |>
  mutate_all(~ log10(. + 1)) |>
  mutate_all(~ (. - mean(.)) / sd(.))
```

```{r}
# Define the predictor variables
X <- as.matrix(dat[c(1:32)])
Y <- as.matrix(dat[c(33:35)])

# group index for X variables
group<- as.integer(c(rep(1,times=3), rep(2,times=6), rep(3, times=3),
                     rep(4,times=4), rep(5, times=3), rep(6,times=2),
                     rep(7,times=6), rep(8, times=2), rep(9,times=3)))
```

```{r}
pacman::p_load(gglasso, grpreg, sparsegl, Matrix, glmnet, stabs)
```

# Loop all three outcomes for feature selection using different group lasso models
```{r}
var_selec <- function(X, Y, group) {
  results <- list()
  for (i in 1:3) {
    y <- Y[, i]
    # Group graphical Lasso
    gr_cv <- cv.gglasso(X, y, group=group, loss="ls", pred.loss="L2",  nfolds=10)
    gr_min_beta <- coef(gr_cv, s = gr_cv$lambda.min)[-1]
  
    # Group Lasso
    grpp_cv <- cv.grpreg(X, y, group = group, penalty="grLasso",seed=5678,nfolds = 10)
    grpp_min_beta <- coef(grpp_cv, s = grpp_cv$lambda.min)[-1]
  
    #Sparse lasso
    sparse_cv<- cv.sparsegl(X, y, group = group, family = "gaussian", nfolds = 10)
    sparse_min_beta<- coef(sparse_cv, s= sparse_cv$lambda.min)[-1]


    #Stability selection with error control - input cross-validated lambda.min from cv-glmnet
    stab_lambda_min <- cv.glmnet(X, y, nfolds=10)$lambda.min
    stab_maxCoef <- stabsel(X, y, fitfun = glmnet.lasso_maxCoef, args.fitfun = list(lambda = stab_lambda_min), cutoff = 0.75, PFER = 1)
    stab_maxCoef_selec<- stab_maxCoef$max
  
    # Store results in list
    results[[paste0("outcome", i)]] <- as.data.frame(list(gr_lasso = gr_min_beta, 
                                                          grpp_lasso = grpp_min_beta,
                                                          sparse_lasso = sparse_min_beta,
                                                          stab_cv_glmnet = stab_maxCoef_selec))
  }
  # Return list of results
  return(results)
}
```


```{r}
res<- var_selec(X, Y, group)
```

#compile results into a dataframe
```{r}
# Add group variable to each list
res <- lapply(res, function(x) {
  df <- as.data.frame(x)
  df$group <- group
  round(df, 5)
})


combined_df_fin<- bind_rows(res, .id = "list_name") %>%
  mutate(list_name = fct_recode(list_name, FSIQ = "outcome1", 
                                VIQ = "outcome2", PIQ = "outcome3")) |>
  rownames_to_column(var = "row_name") |>
  mutate(row_name = str_remove_all(row_name, "\\.\\.\\.\\d+"))


#expot data for manually cleaning the variable names
write_csv(combined_df_fin, paste0(res, "varsel.csv"))

```

#visualize group lasso results
```{r}

#use the modified data 
combined_df_fin<- read_csv(paste0(res, "varsel.csv")) |>
  pivot_longer(!c(row_name, group, list_name), names_to = "model", values_to = "coef") |>
  filter(model=="grpp_lasso" & group != 9)|>
  

ggplot(combined_df_fin, aes(x = model, y = row_name, 
                            fill = ifelse(coef == 0 & !is.na(coef), "gray80", 
                                          ifelse(coef > 0, "royalblue1", "orangered1"))))+
  geom_tile(color="gray50", width=0.5) + 
  scale_fill_identity() +
  theme_minimal() + 
  labs(x = "", y = "", fill = "Coefficient") +  
  facet_grid(group~list_name, scales = "free_y", space = "free_y", switch = "both") + 
  geom_text(aes(label = ifelse(!is.na(coef), round(coef, 4), "")), 
            color = "white", size = 4, fontface = "bold", show.legend = FALSE)+
  scale_x_discrete(expand = c(0, 0))+
  theme(axis.text=element_text(size=11), 
        axis.title=element_text(size=11,face="bold"),
        strip.text = element_text(size=11),
        legend.text=element_text(size=11),
        legend.title=element_text(size=11),
        legend.position="bottom",
        legend.justification="right",
        legend.box="horizontal",
        legend.box.just="center",
        legend.margin=margin(t=0.1, r=0.1, b=2, l=0.1),
        panel.grid.minor = element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.major.y = element_blank(),
        axis.text.x = element_blank(),
        axis.text.y = element_text(size = 11), 
        panel.spacing.x=unit(0.8, "lines"),
        panel.spacing.y=unit(0.02, "lines"),
        strip.text.y = element_blank())
  
```

```{r}
ggsave("E:/BBK17/pj/data_2023apr/results/var_sel/varsel.tiff", 
       width = 4,height = 9,
       dpi=300)
```
