==> 04_bmim_test.Rmd <==
---
title: "bmim_test"
author: "Puvvula"
date: "2023-01-10"
output: pdf_document
---

```{r}
library(bsmim2)
library(bkmr)
library(GIGrvg)
library(refund)
library(ggplot2)
library(gridExtra)
library(tidyverse)
library(splines)
library(MASS)
#library(qgcomp)
```


```{r}
runLocal=TRUE

## params
R <- 120000            ## no. of iterations
burn <- 0.40            ## percent burn-in
thin <- 20             ## thinning number
doLog <- FALSE         ## dont log transform the exposures
# swapX <- TRUE          ## permute exposures (within groups) to avoid artifacts as a result of unique correlation structure
# folds <- 4             ## no. of folds for CV
jump <- 0.35           ## sd of random walk for theta* bsmim
sel <- seq(burn*R+1,R,by=thin) 
dat_names <- c("index3","all") ## different index settings
mod_names <- c("unconstrained","constrained","ordered","dirichlet","dirichlet_varsel","TEQ") ## different models to fit
prior_b0 <- 0.5 ## controls amount of variable selection: Beta(1,prior_b0)


nhanes <- na.omit(read.csv("E:/BBK17/pj//bmim_test/studypop.csv"))

```

Call functions from GitHub
```{r}
source(paste0("https://raw.githubusercontent.com/glenmcgee/infpriormixtures/main/NHANES_analysis/NHANES_cleandat_infprior.R"))
```

Function to generate confidence intervals
```{r}
printCI <- function(df,col1,col2,dig=2){
  
  CI <- paste0("(",round(df[,col1],dig),", ",round(df[,col2],dig),")")
  res <- CI
  if(col1>1){
    res <- cbind(df[1:(col1-1)],CI)
  }
  if(col2<ncol(df)){
    res <- cbind(res,df[(col2+1):ncol(df)])
  }
  return(res)
}
```


```{r}
dat<- prep_data_full()
y<-dat$y
```


```{r}
################################################
###             Format Exposures             ###
################################################
X_list <- dat$bsmim$X  ## all 3 indices (18 exposures)
cnstrs <- c(0,0)
```


Constrained 3 index model 
(https://github.com/glenmcgee/infpriormixtures/blob/main/NHANES_analysis/NHANES_infprior_analysis.R
line 279)
```{r}
fit <- bsmim2(y=y,x=X_list,z=dat$covariates,niter=R,nburn=R*burn,nthin=thin,prior_sigma=c(0.001,0.001),prior_lambda_shaperate=c(1,0.1),gaussian=TRUE,spike_slab=TRUE,gauss_prior=TRUE,prior_theta_slab_sd=0.25,stepsize_theta=jump,basis.opts=NULL,draw_h=FALSE,constraints=c(cnstrs,1),prior_slabpos=c(1.6,8),prior_pi=c(1,prior_b0)) ## shape and rate for gamma on thetastar
```



```{r}
pred_assoc <- predict_hnew_assoc2(fit)
pred_overall <- predict_hnew_assoc2(fit,overall = TRUE)

pred_ind<- predict_hnew_indexwise2(fit)
pred_inter<- pred_twoway(fit)

cv <- bsmim_crossval2(fit,kfolds=4) 

res_list <- list(fit=fit,y=y,pred_assoc=pred_assoc,pred_overall=pred_overall,pred_ind=pred_ind,pred_inter=pred_inter)
save(res_list, file="E://BBK17//pj//bkmr//bmim_test.RData") #load this for results
```


```{r}
postdf_constrained <- data.frame(res_list$fit$thetaPOS[[3]]); colnames(postdf_constrained) <- as.character(1:8)
postdf_constrained <- gather(postdf_constrained,component,weight)

#Weight of each chemical corresponding to joint association
weights_bsmim <- rbind(summarize_thetas(fit)[[1]],summarize_thetas(fit)[[2]],
                       summarize_thetas(fit)[[3]])
weights_bsmim <- printCI(round(weights_bsmim,2),4,5)

#Overall effect estimate
overall_bsmim <- pred_overall$contrasts[12,]
```

Plotting results
```{r}
#Index wise exposure-response association curves
pp_bsmim_ind <- plot_univar_hnew_indexwise2(pred_ind)

#indexwise interactions
ylim <- c(-0.5,0.75)
pp_bsmim <- plot_univar_hnew2(pred_assoc,assoc=F,ylims=ylim)

pp_inter_bsmim <- ggplot(pred_inter, aes(grid, est)) + 
  geom_smooth(aes(col = as.factor(quantile)), stat = "identity",fill="white") + 
  scale_colour_ghibli_d("LaputaMedium",-1)+
  # scale_color_manual(values=wes_palette(n=3, name="GrandBudapest"))+
  facet_grid(var2 ~ var1, scales = "free_x") + ## free_x allows different x axis limits
  ggtitle("Indexwise Interactions") + 
  labs(x="",y="",col="Quantile")+
  theme(strip.text = element_text(colour = 'black'))+
  theme_light()

```
