==> 05.01_linear_reg.Rmd <==
---
title: "05.01_linear_reg"
author: "Jagadeesh Puvvula"
date: "2023-05-12"
output: pdf_document
---

```{r}
library(pacman)
pacman::p_load(tidyverse, janitor)
path <- "E:/BBK17/pj/data_2023apr/obs_w_outcome/"

dat<- read_csv(paste0(path, "wppsi_f_analy.csv"))
```

```{r, echo=FALSE, message=FALSE}
dat<- dat |>
  mutate_if(is.numeric, round, 4) |>
  clean_names() |>
  rename(Pb=pb, Hg=hg, DMA=dma, DDE= dde, PBDE_47 = pbde47,
         PCB_118=pcb118, PCB_138=pcb138, PCB_153=pcb153,
         PCB_180=pcb180, PFHxS=pfhxs, PFOA=pfoa, PFOS=pfos,
         BCEP=bcep, BDCIPP=bdcipp, DNBP=dnbp, DPhP=dphp,
         TCS=tcs, BPA=bpa, MBP=mbp, MBZP=mbzp, MCPP=mcpp,
         sigma_DEHP=dehp, MEP=mep, MIBP=mibp, di_Ethyl_OP=op_de,
         di_Methyl_OP=op_dm, B_PB=b_pb, M_PB=m_pb, P_PB=p_pb, city=center,
         mom_edu_cat=edu3)|>
  mutate(race_bin = ifelse(race == "white", 1, 2))|>
  mutate(sex = ifelse(sex == "1", "Male", "Female"))|>
  select(c(Pb, Hg, DMA, DDE, PBDE_47, PCB_118, PCB_138, PCB_153, PCB_180,
          PFHxS, PFOA, PFOS, BCEP, BDCIPP, DNBP, DPhP, TCS, BPA, MBP, MBZP,
          MCPP, sigma_DEHP, MEP, MIBP, di_Ethyl_OP, di_Methyl_OP, 
          B_PB, M_PB, P_PB, cotinine, home_score_total, mom_age, 
          cohort, city, sex, race_bin,  mom_edu_cat, parity_n, fish_int_pm,
          wppsi_fsiq, wppsi_viq, wppsi_piq))

dat<- dat |> 
  mutate(across(all_of(c("cohort", "city", "sex", "race_bin", 
                         "parity_n", "mom_edu_cat", "fish_int_pm")), as.factor))
```

for 29 chemicals
```{r}
# for all chemicals select 1:32
dat <- dat |>
  select(1:42) |>
  drop_na()|>
  mutate_at(vars(1:30), ~log10(.+0.0000001)) |>
  rename_with(~ paste0("log_", .x), 1:30) |>
  drop_na()
```

#standardized data to have mean =0 SD =1
```{r}
exclude_cols <- 33:39  # Define columns to exclude
selected_cols <- setdiff(seq_along(dat), exclude_cols)  # Columns to standardize

std_dat <- dat
std_dat[, selected_cols] <- Make.Z(dat[, selected_cols])
```



#Get lm results using the function
```{r}
lm_reas<- lm_func(dependent_vars = c("wppsi_fsiq", "wppsi_viq", "wppsi_piq"),
            independent_vars=c("log_Pb", "log_Hg", "log_DMA", "log_DDE", "log_PBDE_47",
                               "log_PCB_118", "log_PCB_138", "log_PCB_153", "log_PCB_180", 
                               "log_PFHxS", "log_PFOA", "log_PFOS", "log_BCEtP", "log_BDCPP",
                               "log_DBuP", "log_DPhP", "log_TCS", "log_BPA", "log_MBP",
                               "log_MBZP", "log_MCPP", "log_sigma_DEHP", "log_MEP", "log_MIBP",
                               "log_di_Ethyl_OP", "log_di_Methyl_OP", "log_B_PB", "log_M_PB", "log_P_PB"),
            covariates = c("log_cotinine", "home_score_total", "mom_age",
                           "race_bin", "mom_edu_cat", "parity_n", "fish_int_pm"),
            data = dat, include_sex = TRUE, include_cohort = TRUE,
            conf_level = 0.80)
```

# Save lm results
```{r}
write_csv(lm_reas, "E:/BBK17/pj/data_2023apr/results/lm/lm_single_chem_stdData_ci80.csv")

```

#data prep for visualization

```{r}
lm_reas<- read_csv("E:/BBK17/pj/data_2023apr/results/lm/lm_single_chem.csv")

dat_lm_res<- lm_reas |>
  rename(chemical="independent_variable", outcome="dependent_variable")|>
  mutate(p_sig= if_else(p_value<0.05, "sig", "non_sig"),
         ci_null = (ci_lower <=0 & ci_upper >=0),
         chemical=gsub("log_", "", chemical),
         cohort_level = fct_recode(cohort_level, "Pooled" = "all",
                             "HOME" = "home",
                             "MIREC" = "mirec"),
         sex_level =fct_recode(sex_level, "All"= "all"),
         outcome =fct_recode(outcome, "FSIQ"= "wppsi_fsiq",
                             "VIQ" = "wppsi_viq",
                             "PIQ" = "wppsi_piq"))

#format chemical name appearance
dat_lm_res<- dat_lm_res |>
  mutate(chemical = toupper(str_replace_all(chemical, "_(?=[A-Z])", " ")),
         chemical = str_replace_all(chemical, "_(?=[0-9])", "")) |>
  mutate(chemical = as.character(chemical))

```


```{r}
#overlapping outcomes
dat_lm_res <- dat_lm_res |>
  group_by(outcome, chemical, sex_level) |>
  mutate(coh_olap = ifelse(all(cohort_level != "Pooled") && (all(coefficient >= 0) | all(coefficient <= 0)), 1, 0)) |>
  ungroup() |>
  group_by(outcome, chemical, cohort_level) |>
  mutate(sex_olap = ifelse(all(sex_level != "All") && (all(coefficient >= 0) | all(coefficient <= 0)), 1, 0)) |>
  ungroup()

# specify the order of levels in a factor variable
dat_lm_res$chemical <- factor(dat_lm_res$chemical,
                      levels = c("PB", "HG", "DMA", "DDE", "PBDE47", "PCB118", "PCB138", "PCB153", "PCB180",
                                 "PFHXS", "PFOA", "PFOS", "BCEP", "BDCIPP", "DNBP", "DPHP", "TCS", "BPA", "MBP",
                                 "MBZP", "MCPP", "SIGMA DEHP", "MEP", "MIBP", "DI ETHYL OP", "DI METHYL OP",
                                 "B PB", "M PB", "P PB"))
```

#filter only pooled results
```{r}
dat_lm_res <- dat_lm_res |>
  filter(cohort_level == "Pooled")
```



# visualization of lm results
```{r}
ggplot(dat_lm_res, aes(x = sex_level, y = chemical)) +
  geom_tile(aes(fill = coefficient)) + #, alpha = ifelse(ci_null == "FALSE", 1, 0) #to remove fill color if ns
  geom_text(aes(label = round(coefficient, 2),
                fontface = ifelse(ci_null == "FALSE", "bold", "plain")),
            color = "black", size = 3, show.legend = FALSE) +
  labs(x = "", y = "", fill = "coefficient") +
  facet_grid(.~ outcome, scales = "free_y", space = "free_y") +
  theme_bw() +
  theme(axis.text = element_text(size = 10), 
        axis.title = element_text(size = 10, face = "bold"),
        panel.grid.minor = element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.major.y = element_blank(),
        axis.text.x = element_text(angle = 0, hjust = 0.5, vjust = 1, size = 10),
        axis.text.y = element_text(size = 10), 
        panel.spacing.x = unit(0.02, "lines"),
        panel.spacing.y = unit(0.02, "lines"),
        strip.text.y = element_blank(),
        strip.text.x = element_text(face= "bold", size = 10),
        strip.background.x = element_rect(fill =  "transparent"),
        legend.position = "bottom", legend.box = "horizontal",
        legend.key.width = unit(0.8, "cm"),
        legend.key.height = unit(0.3, "cm"),
        legend.spacing = unit(0, "cm")) +
  scale_fill_gradient2(low = "red", high = "royalblue", 
                       midpoint = 0, name = expression(beta[italic(lm)]),
                       guide = guide_colorbar(title.position = "left",
                                              title.hjust = 1,
                                              title.vjust = 1)) +
  scale_alpha_continuous(range = c(0, 1), guide = "none")  # Remove alpha legend

```

#export lm visualization
```{r}
ggsave("E:/BBK17/pj/data_2023apr/results/lm/lm_coef.tiff", 
       width = 7,height = 6,
       dpi=300)
```

#======================================================#
#converting lm results to table - suggested by Michael
```{r}
lm_res_tbl <- dat_lm_res |>
  filter(cohort_level == "Pooled") |>
  mutate(result = sprintf("%.2f (%.2f, %.2f)", coefficient, ci_lower, ci_upper)) |>
  select(-c(4:10)) |>
  pivot_wider(names_from = c(outcome, sex_level),
              values_from = result ) |>
  select(
    chemical,
    FSIQ_All, FSIQ_Female, FSIQ_Male,  
    VIQ_All, VIQ_Female, VIQ_Male,
    PIQ_All, PIQ_Female, PIQ_Male)

write_csv(lm_res_tbl, "E:/BBK17/pj/data_2023apr/results/lm/lm_res_fmt.csv")
```
