==> 06.01_chem_conc_grp_diff.Rmd <==
---
title: "06.01_chem_conc_grp_diff"
author: "Jagadeesh Puvvula"
date: "2023-07-20"
output: pdf_document
---

```{r, echo=FALSE, message=FALSE}

library(pacman)
pacman::p_load(tidyverse, janitor, bkmr, Matrix, fields, bindrcpp)
path <- "E:/BBK17/pj/data_2023apr/obs_w_outcome/"

dat<- read_csv(paste0(path, "wppsi_f_analy.csv"))

```

```{r, echo=FALSE, message=FALSE}
dat<- dat |>
  mutate_if(is.numeric, round, 4) |>
  clean_names() |>
  mutate(pb = if_else(cohort == "1", pb * 10, pb))|>
  rename(Pb=pb, Hg=hg, DMA=dma, DDE= dde, PBDE_47 = pbde47,
         PCB_118=pcb118, PCB_138=pcb138, PCB_153=pcb153,
         PCB_180=pcb180, PFHxS=pfhxs, PFOA=pfoa, PFOS=pfos,
         BCEtP=bcetp, BDCPP=bdcpp, DBuP=dbup, DPhP=dphp,
         TCS=tcs, BPA=bpa, MBP=mbp, MBZP=mbzp, MCPP=mcpp,
         sigma_DEHP=dehp, MEP=mep, MIBP=mibp, di_Ethyl_OP=op_de,
         di_Methyl_OP=op_dm, B_PB=b_pb, M_PB=m_pb, P_PB=p_pb, city=center,
         mom_edu_cat=edu3)|>
  mutate(race_bin = ifelse(race == "white", 1, 2))|>
  select(c(Pb, Hg, DMA, DDE, PBDE_47, PCB_118, PCB_138, PCB_153, PCB_180,
          PFHxS, PFOA, PFOS, BCEtP, BDCPP, DBuP, DPhP, TCS, BPA, MBP, MBZP,
          MCPP, sigma_DEHP, MEP, MIBP, di_Ethyl_OP, di_Methyl_OP, 
          B_PB, M_PB, P_PB, cotinine, home_score_total, mom_age, 
          cohort, city, sex, race_bin,  mom_edu_cat, parity_n, 
           wppsi_fsiq, wppsi_viq, wppsi_piq)) |>
  drop_na()|>
  mutate_at(vars(1:30), ~log10(.+0.0000001))
```


#linear regression
```{r}
group_diff_lm <- function(data, continuous_vars, group_var, folder_location) {
  
  # Create a new dummy variable for cohort groups
  data$cohort_dummy <- ifelse(data[[group_var]] == "2", 0, 1) #MIREC as reference group
  
  # Perform linear regression for each continuous variable against the cohort_dummy
  results <- lapply(continuous_vars, function(var) {
    model <- lm(get(var) ~ cohort_dummy, data = data)
    
    # Round the estimates, confidence intervals, and p-values
    coef_data <- data.frame(variable = var,
                            beta = round(coef(model)[2], 2),
                            CI = paste(round(confint(model)[2, ], 2), collapse = ','),
                            p_value = formatC(summary(model)$coefficients[2, 4], digits = 4, format = "f"))
    return(coef_data)
  })
  
  # Combine the results into a single dataframe
  result_df <- do.call(rbind, results)
  
  # Save the result dataframe to a CSV file in the specified folder location
  output_file <- file.path(folder_location, "linear_regression_results.csv")
  write.csv(result_df, file = output_file, row.names = FALSE)
  
  return(result_df)
}
```

```{r}
group_diff_lm(data=dat,
              continuous_vars = colnames(dat)[c(1:32, 39:41)],
              group_var = "cohort",
              folder_location = "E:/BBK17/pj/data_2023apr/results/grp_diff")
```
