==> 07_supplement.Rmd <==
---
title: "07_supplement"
author: "Jagadeesh Puvvula"
date: "2023-12-11"
output: pdf_document
---

#load env
```{r}
library(pacman)
pacman::p_load(tidyverse, janitor, qgcomp)
path <- "E:/BBK17/pj/data_2023apr/obs_w_outcome/"
```

#import and format data
```{r, echo=FALSE, message=FALSE}
dat<- read_csv(paste0(path, "wppsi_f_analy.csv")) |>
  mutate_if(is.numeric, round, 4) |>
  clean_names() |>
  rename(Pb=pb, Hg=hg, DMA=dma, DDE= dde, PBDE_47 = pbde47,
         PCB_118=pcb118, PCB_138=pcb138, PCB_153=pcb153,
         PCB_180=pcb180, PFHxS=pfhxs, PFOA=pfoa, PFOS=pfos,
         BCEP=bcep, BDCIPP=bdcipp, DNBP=dnbp, DPhP=dphp,
         TCS=tcs, BPA=bpa, MBP=mbp, MBZP=mbzp, MCPP=mcpp,
         sigma_DEHP=dehp, MEP=mep, MIBP=mibp, di_Ethyl_OP=op_de,
         di_Methyl_OP=op_dm, B_PB=b_pb, M_PB=m_pb, P_PB=p_pb, city=center,
         mom_edu_cat=edu3)|>
  mutate(race_bin = ifelse(race == "white", 1, 2))|>
  mutate(sex = ifelse(sex == "1", "Male", "Female"))|>
  select(c(Pb, Hg, DMA, DDE, PBDE_47, PCB_118, PCB_138, PCB_153, PCB_180,
          PFHxS, PFOA, PFOS, BCEP, BDCIPP, DNBP, DPhP, TCS, BPA, MBP, MBZP,
          MCPP, sigma_DEHP, MEP, MIBP, di_Ethyl_OP, di_Methyl_OP, 
          B_PB, M_PB, P_PB, cotinine, home_score_total, mom_age, 
          cohort, city, sex, race_bin,  mom_edu_cat, parity_n, fish_int_pm,
          wppsi_fsiq, wppsi_viq, wppsi_piq))

dat<- dat |> 
  mutate(across(all_of(c("cohort", "city", "sex", "race_bin", 
                         "parity_n", "mom_edu_cat", "fish_int_pm")), as.factor))
```

#piq - male - 9 chemical biomarkers - n=333
```{r}
piq_male<- dat |>
  select( "Hg", "PBDE_47", "PFHxS", "PFOA", "PFOS", "BDCIPP", "MBP", "MCPP", "P_PB",
          "cotinine", "race_bin", "mom_edu_cat", "home_score_total", "parity_n", "mom_age", "fish_int_pm",
          "city", "cohort", "sex",
          "wppsi_piq") |>
  drop_na()|>
  mutate_at(vars(1:10), ~log10(.+0.0000001)) |>
  rename_with(~ paste0("log_", .x), 1:10) |>
  drop_na() 

summary(factor(piq_male$sex))

mixture<- names(piq_male[c(1:9)])

qgcomp_func(outcomes = c("wppsi_piq"), 
            data = piq_male,
            output_folder = "E:/BBK17/pj/data_2023apr/results/qgcomp_supplement", 
            include_sex = TRUE, 
            include_cohort = FALSE, 
            chemicals = list("log_Hg", "log_PBDE_47", "log_PFHxS", "log_PFOA", 
                             "log_PFOS", "log_BDCIPP", "log_MBP", "log_MCPP", "log_P_PB"),
            covariates = list("race_bin", "log_cotinine", "mom_edu_cat", 
                              "home_score_total", "parity_n", "mom_age", 
                              "fish_int_pm"),
            q = 10, b=400, seed = 2022)
```

#piq - female - 5 chemical biomarkers - n=388
```{r}
piq_female<- dat |>
  select( "PCB_138", "PFHxS", "PFOA", "PFOS", "BCEP",
          "cotinine", "race_bin", "mom_edu_cat", "home_score_total", "parity_n", "mom_age", "fish_int_pm",
          "city", "cohort", "sex",
          "wppsi_piq") |>
  drop_na()|>
  mutate_at(vars(1:6), ~log10(.+0.0000001)) |>
  rename_with(~ paste0("log_", .x), 1:6) |>
  drop_na() 

mixture<- names(piq_female[c(1:5)])

qgcomp_func(outcomes = c("wppsi_piq"), 
            data = piq_female,
            output_folder = "E:/BBK17/pj/data_2023apr/results/qgcomp_supplement", 
            include_sex = TRUE, 
            include_cohort = FALSE, 
            chemicals = list("log_PCB_138",  "log_PFHxS", "log_PFOA", "log_PFOS", "log_BCEP"),
            covariates = list("race_bin", "log_cotinine", "mom_edu_cat", 
                              "home_score_total", "parity_n", "mom_age", 
                              "fish_int_pm"),
            q = 10, b=400, seed = 2022)
```


#piq - pooled - 5 chemical biomarkers - n=728
```{r}
piq_pooled<- dat |>
  select("Hg", "PBDE_47", "PCB_118", "PCB_138", "MCPP", 
         "cotinine", "race_bin", "mom_edu_cat", 
         "home_score_total", "parity_n", "mom_age", "fish_int_pm", "wppsi_piq",
         "city", "cohort", "sex") |>
  drop_na()|>
  mutate_at(vars(1:6), ~log10(.+0.0000001)) |>
  rename_with(~ paste0("log_", .x), 1:6) |>
  drop_na()

mixture<- names(piq_pooled[c(1:5)])

qgcomp_func(outcomes = c("wppsi_piq"), 
            data = piq_pooled,
            output_folder = "E:/BBK17/pj/data_2023apr/results/qgcomp_supplement", 
            include_sex = FALSE, 
            include_cohort = FALSE, 
            chemicals = list("log_Hg", "log_PBDE_47", "log_PCB_118", "log_PCB_138", "log_MCPP"),
            covariates = list("race_bin", "log_cotinine", "mom_edu_cat", 
                              "home_score_total", "parity_n", "mom_age", 
                              "fish_int_pm"),
            q = 10, b=400, seed = 2022)
```


#fsiq - pooled - 6 chemical biomarkers - n=718
```{r}
fsiq_pooled<- dat |>
  select("PBDE_47", "PCB_118", "PCB_138", "PFOA", "MCPP", "B_PB",
         "cotinine", "race_bin", "mom_edu_cat", 
         "home_score_total", "parity_n", "mom_age", "fish_int_pm", "wppsi_fsiq",
         "city", "cohort", "sex") |>
  drop_na()|>
  mutate_at(vars(1:7), ~log10(.+0.0000001)) |>
  rename_with(~ paste0("log_", .x), 1:7) |>
  drop_na()

mixture<- names(fsiq_pooled[c(1:6)])

qgcomp_func(outcomes = c("wppsi_fsiq"), 
            data = fsiq_pooled,
            output_folder = "E:/BBK17/pj/data_2023apr/results/qgcomp_supplement", 
            include_sex = FALSE, 
            include_cohort = FALSE, 
            chemicals = list("log_PBDE_47", "log_PCB_118", "log_PCB_138", "log_PFOA", "log_MCPP", "log_B_PB"),
            covariates = list("race_bin", "log_cotinine", "mom_edu_cat", 
                              "home_score_total", "parity_n", "mom_age", 
                              "fish_int_pm"),
            q = 10, b=400, seed = 2022)
```

#fsiq - male - 5 chemical biomarkers - n=346
```{r}
fsiq_male<- dat |>
  select("PBDE_47", "PFHxS", "MCPP", "di_Ethyl_OP", "B_PB",
         "cotinine", "race_bin", "mom_edu_cat", 
         "home_score_total", "parity_n", "mom_age", "fish_int_pm", "wppsi_fsiq",
         "city", "cohort", "sex") |>
  drop_na()|>
  mutate_at(vars(1:6), ~log10(.+0.0000001)) |>
  rename_with(~ paste0("log_", .x), 1:6) |>
  drop_na()

summary(factor(fsiq_male$sex))

mixture<- names(fsiq_male[c(1:5)])

qgcomp_func(outcomes = c("wppsi_fsiq"), 
            data = fsiq_male,
            output_folder = "E:/BBK17/pj/data_2023apr/results/qgcomp_supplement", 
            include_sex = T, 
            include_cohort = FALSE, 
            chemicals = list("log_PBDE_47", "log_PFHxS", "log_MCPP", "log_di_Ethyl_OP", "log_B_PB"),
            covariates = list("race_bin", "log_cotinine", "mom_edu_cat", 
                              "home_score_total", "parity_n", "mom_age", 
                              "fish_int_pm"),
            q = 10, b=400, seed = 2022)
```


#fsiq - female - 6 chemical biomarkers - n=382
```{r}
fsiq_female<- dat |>
  select("PCB_138", "PCB_153", "PFHxS", "PFOA", "PFOS", "MEP",
         "cotinine", "race_bin", "mom_edu_cat", 
         "home_score_total", "parity_n", "mom_age", "fish_int_pm", "wppsi_fsiq",
         "city", "cohort", "sex") |>
  drop_na()|>
  mutate_at(vars(1:7), ~log10(.+0.0000001)) |>
  rename_with(~ paste0("log_", .x), 1:7) |>
  drop_na()

summary(factor(fsiq_female$sex))

mixture<- names(fsiq_female[c(1:6)])

qgcomp_func(outcomes = c("wppsi_fsiq"), 
            data = fsiq_female,
            output_folder = "E:/BBK17/pj/data_2023apr/results/qgcomp_supplement", 
            include_sex = T, 
            include_cohort = FALSE, 
            chemicals = list("log_PCB_138", "log_PCB_153", "log_PFHxS", "log_PFOA", "log_PFOS", "log_MEP"),
            covariates = list("race_bin", "log_cotinine", "mom_edu_cat", 
                              "home_score_total", "parity_n", "mom_age", 
                              "fish_int_pm"),
            q = 10, b=400, seed = 2022)
```


#viq - pooled - 8 chemical biomarkers - n=738
```{r}
viq_pooled<- dat |>
  select("PBDE_47", "PCB_118", "PCB_138", "PCB_153", "PFOA", "PFOS", "BPA", "B_PB",
         "cotinine", "race_bin", "mom_edu_cat", 
         "home_score_total", "parity_n", "mom_age", "fish_int_pm", "wppsi_viq",
         "city", "cohort", "sex") |>
  drop_na()|>
  mutate_at(vars(1:9), ~log10(.+0.0000001)) |>
  rename_with(~ paste0("log_", .x), 1:9) |>
  drop_na()

summary(factor(fsiq_female$sex))

mixture<- names(viq_pooled[c(1:8)])

qgcomp_func(outcomes = c("wppsi_viq"), 
            data = viq_pooled,
            output_folder = "E:/BBK17/pj/data_2023apr/results/qgcomp_supplement", 
            include_sex = FALSE, 
            include_cohort = FALSE, 
            chemicals = list("log_PBDE_47", "log_PCB_118", "log_PCB_138", "log_PCB_153",
                             "log_PFOA", "log_PFOS", "log_BPA", "log_B_PB"),
            covariates = list("race_bin", "log_cotinine", "mom_edu_cat", 
                              "home_score_total", "parity_n", "mom_age", 
                              "fish_int_pm"),
            q = 10, b=400, seed = 2022)
```


#viq - male - 2 chemical biomarkers - n=370
```{r}
viq_male<- dat |>
  select("PBDE_47", "B_PB",
         "cotinine", "race_bin", "mom_edu_cat", 
         "home_score_total", "parity_n", "mom_age", "fish_int_pm", "wppsi_viq",
         "city", "cohort", "sex") |>
  drop_na()|>
  mutate_at(vars(1:3), ~log10(.+0.0000001)) |>
  rename_with(~ paste0("log_", .x), 1:3) |>
  drop_na()

summary(factor(viq_male$sex))

mixture<- names(viq_male[c(1:2)])

qgcomp_func(outcomes = c("wppsi_viq"), 
            data = viq_male,
            output_folder = "E:/BBK17/pj/data_2023apr/results/qgcomp_supplement", 
            include_sex = TRUE, 
            include_cohort = FALSE, 
            chemicals = list("log_PBDE_47", "log_B_PB"),
            covariates = list("race_bin", "log_cotinine", "mom_edu_cat", 
                              "home_score_total", "parity_n", "mom_age", 
                              "fish_int_pm"),
            q = 10, b=400, seed = 2022)
```

#viq - female - 2 chemical biomarkers - n=396
```{r}
viq_female<- dat |>
  select("PCB_138", "BPA", "MEP",
         "cotinine", "race_bin", "mom_edu_cat", 
         "home_score_total", "parity_n", "mom_age", "fish_int_pm", "wppsi_viq",
         "city", "cohort", "sex") |>
  drop_na()|>
  mutate_at(vars(1:4), ~log10(.+0.0000001)) |>
  rename_with(~ paste0("log_", .x), 1:4) |>
  drop_na()

summary(factor(viq_female$sex))

mixture<- names(viq_female[c(1:3)])

qgcomp_func(outcomes = c("wppsi_viq"), 
            data = viq_female,
            output_folder = "E:/BBK17/pj/data_2023apr/results/qgcomp_supplement", 
            include_sex = TRUE, 
            include_cohort = FALSE, 
            chemicals = list("log_PCB_138", "log_BPA", "log_MEP"),
            covariates = list("race_bin", "log_cotinine", "mom_edu_cat", 
                              "home_score_total", "parity_n", "mom_age", 
                              "fish_int_pm"),
            q = 10, b=400, seed = 2022)
```

#==============================================================================#
#==============================================================================#

#extract results
```{r}
results<- get_gcomp_estimates(results_folder = "E:/BBK17/pj/data_2023apr/results/qgcomp/qgcomp_subset")
write_csv(results, "E:/BBK17/pj/data_2023apr/results/qgcomp/qgcomp_subset/r_obj_summ.csv")
```

#data prep for visualizing estimates and CI
```{r}
res<- results |>
  rename(bootstrap=boot_strp) |>
  mutate(p_value = if_else(p_value <= 0.05, "<0.05", ">0.05"),
         ci_null = (lower_ci <=0 & upper_ci >=0)) |>
  mutate(cohort = fct_recode(as.factor(cohort), Pooled = "all", HOME = "home", MIREC = "mirec")) |>
  mutate(outcome = fct_recode(as.factor(outcome), FSIQ = "fsiq", VIQ = "viq", PIQ = "piq")) |>
  mutate(gender = fct_recode(as.factor(gender), All = "all", Female = "Female", Male = "Male")) |>
  mutate(bootstrap = fct_recode(as.factor(bootstrap), Yes = "boot", No = "nb")) |>
  filter(abs(lower_ci) <= 40 & abs(upper_ci) <= 40,
         cohort == "Pooled")

#res$cohort<- factor(res$cohort, levels = c("Pooled", "HOME", "MIREC"))
res$bootstrap<- factor(res$bootstrap, levels = c("Yes", "No"))

cbbPalette <- c("black", "#0072B2")
```

#visualizing estimates and CI
```{r, echo=FALSE, message=FALSE, fig.height=4, fig.width=6, dpi=300}
ggplot(res, aes(x = as.factor(outcome), y = estimate, ymin = lower_ci, ymax = upper_ci)) + 
  geom_pointrange(aes(col = as.factor(ci_null), shape=bootstrap), 
                  position=position_dodge(width=0.5),size = 0.6) +
  geom_hline(aes(yintercept = 0), linetype="dashed") + 
  scale_colour_manual(values=cbbPalette) +
  labs(x=NULL,
       y=expression(paste(beta[Q-gcomp]," [95% CI]")),
       title = "",
       caption = "")+
  theme_bw()+
  theme(axis.line = element_line(colour = "black"))+
  theme(plot.caption = element_text(size = 6),
        plot.title=element_text(size = 10, hjust = 0),
        panel.border = element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.line = element_line(colour = "black"),
        text=element_text(size=10),
        axis.text = element_text(size = 10),
        axis.text.x=element_text(size=10,angle=0),
        axis.ticks.x=element_blank(),
        strip.text.x=element_text(size=10,angle=0),
        strip.text.y = element_text(size=10),
        strip.placement = "outside",
        strip.background = element_rect(fill=NULL, colour = NULL),
        legend.box.just = "center",
        legend.spacing = unit(0, "cm"),
        legend.position = "bottom",
        legend.box = "horizontal",
        legend.direction = "horizontal",
        legend.text = element_text(size = 8),
        legend.key = element_blank(),
        legend.key.height = unit(3, "mm"))+
  guides(col="none",
         shape=guide_legend(title = "Bootstrap"))+
  facet_grid(gender~., scales = "free", switch = "x", space = "free_x")
```

#export figure
```{r}
ggsave("E:/BBK17/pj/data_2023apr/results/qgcomp/qgcomp_subset/qgcomp_subset.tiff", 
       width = 4,height = 6,
       dpi=300)
```

#==============================================================================#
#==============================================================================#

#extract chemical weights 
```{r}
qgcomp_weights <- extract_weights("E:/BBK17/pj/data_2023apr/results/qgcomp/qgcomp_subset")
write_csv(qgcomp_weights, "E:/BBK17/pj/data_2023apr/results/qgcomp/qgcomp_subset/chem_weight.csv")
```

#data prep for visualization
```{r}
dat <- qgcomp_weights |> 
  mutate(file_name = str_remove(file_name, "nb_")) |>
  separate(file_name, c("gender", "cohort", "wppsi", "outcome"), sep = "_") |>
  mutate(weight= if_else(direction=="neg", weight*-1, weight),
         chemical=gsub("log_", "", chemical),
         outcome = str_to_upper(outcome),
         cohort = fct_recode(cohort, "Pooled" = "all",
                             "HOME" = "home",
                             "MIREC" = "mirec"),
         gender =fct_recode(gender, "All"= "all"),
         chemical = toupper(str_replace_all(chemical, "_(?=[A-Z])", " ")),
         chemical = str_replace_all(chemical, "_(?=[0-9])", ""),
         chem_grp = ifelse(grepl("psi", direction), "psi", "chem")) |>
  filter(cohort== "Pooled")

# specify the order of levels in a factor variable
dat$chemical <- factor(dat$chemical,
                      levels = c("PB", "HG", "DMA", "DDE", "PBDE47", "PCB118", "PCB138", "PCB153", "PCB180",
                                 "PFHXS", "PFOA", "PFOS", "BCEP", "BDCIPP", "DNBP", "DPHP", "TCS", "BPA", "MBP",
                                 "MBZP", "MCPP", "SIGMA DEHP", "MEP", "MIBP", "DI ETHYL OP", "DI METHYL OP",
                                 "B PB", "M PB", "P PB", "NEG.PSI", "POS.PSI", "PSI1"))
dat$chem_grp<- factor(dat$chem_grp, levels = c("psi", "chem"))
```


#visualize chemical weights
```{r}
ggplot(dat, aes(y = chemical, x = gender, fill = ifelse(chem_grp == "chem", weight, NA))) + 
  geom_tile() +
  geom_tile(data = dat %>% filter(chem_grp == "psi"), fill = "white") +
  geom_text(aes(label = round(weight, 3)),
              color = "black", size = 2.5, fontface = "bold", show.legend = FALSE) +
  labs(x = "", y = "", fill = "weight") +
    facet_grid(chem_grp ~ outcome, scales = "free_y", space = "free_y") +
    theme_bw() +
    theme(axis.text = element_text(size = 10), 
        axis.title = element_text(size = 10, face = "bold"),
        panel.grid.minor = element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.major.y = element_blank(),
        axis.text.x = element_text(angle = 0, hjust = 0.5, vjust = 1, size = 10),
        axis.text.y = element_text(size = 10), 
        panel.spacing.x = unit(0.02, "lines"),
        panel.spacing.y = unit(0.02, "lines"),
        strip.text.y = element_blank(),
        strip.text.x = element_text(face= "bold", size = 10),
        strip.background.x = element_rect(fill =  "transparent"),
        legend.position = "bottom", legend.box = "horizontal",
        legend.key.width = unit(0.8, "cm"),
        legend.key.height = unit(0.3, "cm"),
        legend.spacing = unit(0, "cm")) +
    scale_fill_gradient2(low = "red", high = "royalblue", midpoint = 0,
                         name = "Chemical biomarker weight",
                         guide = guide_colorbar(title.vjust = 1.2),
                         breaks = c(-0.5, 0, 0.5))
```

#export figure
```{r}
ggsave("E:/BBK17/pj/data_2023apr/results/qgcomp/qgcomp_subset/qgcomp_subset_weight.tiff", 
       width = 5.5,height = 6,
       dpi=300)
```
