==> 99_MI_mirec_covars.Rmd <==
---
title: "99_MI_covars_mirec"
author: "Jagadeesh Puvvula"
date: "2023-02-08"
output: pdf_document
---

```{r}
library(tidyverse)
library(caret)
library(SuperLearner)
```


Birth length - predictions
```{r, message=FALSE}
dat <- read_csv("E:/BBK17/pj/wppsi_exp_out_comb/wppsi_home_mirec.csv")
set<- read_csv("E:/BBK17/pj/wppsi_exp_out_comb/b_len_missing_set.csv") |>
  select(c(1))

#revert missing median birth length with NA
dat$b_length <- ifelse(dat$subject_id %in% set$subject_id, NA, dat$b_length)
remove(set)

#filter unknown data for predictions
pred_set<- dat |>
  filter(is.na(b_length)) |>
  select(c(4:37, 40:47))

#for prediction model
dat_n <- dat |> drop_na() |>
  select(c(5:37, 40:47))

set.seed(123, "L'Ecuyer-CMRG")
indx<- sample(nrow(dat_n), round(0.75*nrow(dat_n)))
train<- dat_n[indx,]
test<- dat_n[-indx,]

y_train<- train$b_length
x_train<- train[,-41]

x_test <- as.matrix(test[, -41])
y_test <- as.matrix(test$b_length)


reg.models<- c("SL.mean", "SL.xgboost", "SL.gbm")

sl.obj2<- SuperLearner(y_train, x_train, family = gaussian(), SL.library = reg.models)
sl.obj2

#Predictions for test set
pred_SL_b_len<- predict(sl.obj2, test[-c(41)])

# Print the mean squared error: 2.57
mse<- mean((pred_SL_b_len$pred - test$b_length)^2)
print(paste("Mean Squared Error:", mse))

# Plot the model's predictions versus the actual target values
plot(pred_SL_b_len$pred, test$b_length)+
abline(0, 1)

# PREDICTING UNKNOWN 
b_len_imputed<- predict.SuperLearner(sl.obj2, pred_set[-c(1,42)])
b_len_imputed$pred<- round(b_len_imputed$pred, digits = 2)


save(sl.obj, file = "E:/BBK17/pj/wppsi_results/pred_model/b_len_pred.rda")
```
