==> 99_99_scratch.Rmd <==
```{r}
wqs_func <- function(outcomes, dat, output_folder, covariates,
                     include_sex = TRUE, include_cohort = TRUE, q,
                     b1_pos, b1_constr, rh) {
  
  for (sex_level in c("all", 1, 2)) {
    if (include_sex) {
      if (sex_level == "all") {
        sex_data <- dat
        sex_formula <- "sex + "
        filename_prefix_sex <- "all"
      } else {
        sex_data <- subset(dat, sex == sex_level)
        sex_formula <- ""
        if (sex_level == 1) {
          filename_prefix_sex <- "male"
        } else if (sex_level == 2) {
          filename_prefix_sex <- "female"
        } else {
          filename_prefix_sex <- as.character(sex_level)
        }
      }
    } else {
      sex_data <- dat
      sex_formula <- ""
      filename_prefix_sex <- "all"
    }
    
    for (cohort_level in c("all", 1, 2)) {
      if (include_cohort) {
        if (cohort_level == "all") {
          cohort_data <- sex_data
          cohort_formula <- "cohort + city+"
          filename_prefix_cohort <- "all"
        } else {
          cohort_data <- subset(sex_data, cohort == cohort_level)
          cohort_formula <- ""
          if (cohort_level == 1) {
            filename_prefix_cohort <- "HOME"
          } else if (cohort_level == 2) {
            filename_prefix_cohort <- "MIREC"
          } else {
            filename_prefix_cohort <- as.character(cohort_level)
          }
        }
      } else {
        cohort_data <- sex_data
        cohort_formula <- ""
        filename_prefix_cohort <- "all"
      }
      
      for (outcome in outcomes) {
        # Construct the formula
        formula <- as.formula(paste(outcome, "~ wqs+", sex_formula, cohort_formula, paste(covariates, collapse = " + ")))
        
        tryCatch({
          # Fit the models
          wqs <- gwqs(formula, mix_name = mixture, data = cohort_data, family = "gaussian", q = 4, b = 100, validation = 0.3, seed = 2023, b1_pos = b1_pos, b1_constr = b1_constr)
          wqsrh <- gwqsrh(formula, mix_name = mixture, data = cohort_data, family = "gaussian", q = 4, b = 100, validation = 0.3, seed = 2023, b1_pos = b1_pos, b1_constr = b1_constr, rh = rh)
          
          # Define the filename prefix based on sex and cohort levels
          filename_prefix <- paste0(filename_prefix_sex, "_cohort_", filename_prefix_cohort)
          
          # Save the models
          save(wqs, file = paste0(output_folder, "/", "wqs_", filename_prefix, "_", outcome, ".rda"))
          save(wqsrh, file = paste0(output_folder, "/", "wqsrh_", filename_prefix, "_", outcome, ".rda"))
        }, error = function(e) {
          cat("An error occurred for outcome", outcome, "with sex level", sex_level, "and cohort level", cohort_level, ":\n")
          message("Error message:", e$message, "\n")
          cat("Moving to the next combination.\n")
        })
      }
    }
  }
}


```
